centre_frequency, bandwidth update
centre_frequency assumed to be the centre of the (N/2)th channel
(starting from 0).
bandwidth assumed to be the total analogue bandwidth carried by
channels broadcast but may not be a contiguous band.

The problem is that the 0th channel may carry different frequency
components depending on if the whole band is being broadcast or not.
The frequency band represented by the 0th channel is thus ambiguous
and it should be discarded.